deleted multi-thread test
