--- crawling/img_reader/href_test.py
@@ -1,54 +1,0 @@
-import os.path
-import re
-from bs4 import BeautifulSoup
-import requests
-import ssl
-import os
-import time
-
-ssl._create_default_https_context = ssl._create_unverified_context
-main_url = "https://porterna.com"
-user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36"
-
-def find_last_page(target_url):
-    pattern = r'<li class="xans-record-"><a class="other" href="\?cate_no=\d+&amp;page=(\d+)">'
-    header = {
-        'Referrer': main_url,
-        'user-agent': user_agent
-    }
-    response = requests.get(target_url, headers=header)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    find_li = soup.find('div', 'xans-element- xans-product xans-product-normalpaging paging')
-    find_li = find_li.find_all('li')
-    find_li = list(map(str, find_li))
-    string = str(find_li[-1])
-    match = re.search(pattern, string)
-    number = match.group(1)
-    return number
-
-
-def detail_url_scraper(target_url):
-    href_list = []
-    for page_num in range(1, int(find_last_page(target_url)) + 1):
-        header = {
-            'Referrer': target_url + str(page_num),
-            'user-agent': user_agent
-        }
-        response = requests.get(target_url + str(page_num), headers=header)
-        soup = BeautifulSoup(response.text, 'html.parser')
-        soup = soup.find_all('ul', {'class': 'thumbnail'})
-        href_set = set()
-        for ul_tag in soup:
-            a_tags = ul_tag.find_all('a')
-            for a_tag in a_tags:
-                href = a_tag.get('href')
-                if href is not None:
-                    href_set.add(href)
-        href_list += list(href_set)
-    print(len(href_list))
-
-if __name__ == "__main__":
-    start_time = time.time()
-    target_url = "https://porterna.com/product/list.html?cate_no=79&page="
-    detail_url_scraper(target_url)
-    print(f"{time.time() - start_time} sec")
